pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name="igibson"
version="2.2.2"
readme = "README.md"
requires-python = ">=2.8"
license = {text = "MIT"}
dependencies = [
    "gym>=0.12",
    "numpy>=1.16.0",
    "scipy>=1.4",
    "pybullet-svl>=3.1.6.4",
    "transforms3d>=0.3.1",
    "opencv-python>=3.4.8",
    "Pillow>=5.4.0",
    "networkx>=2.0",
    "PyYAML",
    "tqdm",
    "freetype-py",
    "matplotlib",
    "cloudpickle",
    "aenum",
    "GPUtil",
    "ipython",
    "future",
    "trimesh",
    "h5py",
    "gitpython",
    "py360convert",
    "six",
    "pandas",
    "packaging",
    "bddl~=1.0.1",
    "urllib3>=1.20",
    "progressbar>=2.5",
    "packaging",
    "pyquaternion",
    "fire",
]

[tool.black]
line-length = 120
target-version = ['py36', 'py37', 'py38']
force-exclude = 'igibson/(data|external|render/(cpp|cryptopp|glad|glfw|glm|pybind11|sranipal))'

[tool.isort]
profile = "black"
line_length = 120
py_version = 'all'
filter_files = true
extend_skip_glob = [
    'igibson/data/*',
    'igibson/external/*',
    'igibson/render/cpp/*',
    'igibson/render/cryptopp/*',
    'igibson/render/glad/*',
    'igibson/render/glfw/*',
    'igibson/render/glm/*',
    'igibson/render/pybind11/*',
    'igibson/render/sranipal/*',
]

[tool.pyright]
exclude = [
  'igibson/data',
  'igibson/docs',
  'igibson/docker',
  'igibson/render/cryptopp',
  'igibson/render/cpp',
  'igibson/render/glfw',
  'igibson/render/glad',
  'igibson/render/glm',
  'igibson/render/pybind11',
  'igibson/render/sranipal'
]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]
addopts = '--cov=igibson --cov-report=xml'

[tool.coverage.run]
omit = ["igibson/external/*"]
